+++ tests/test_mapping.py
@@ -3,13 +3,13 @@ from tempfile import TemporaryDirectory
 
 import pytest
 
-from pdm_conda.mapping import DOWNLOAD_DIR_ENV_VAR
+from pdm_conda.mapping import MAPPING_DOWNLOAD_DIR_ENV_VAR
 
 
 @pytest.fixture
 def patch_download_dir(monkeypatch):
     with TemporaryDirectory() as d:
-        monkeypatch.setenv(DOWNLOAD_DIR_ENV_VAR, d)
+        monkeypatch.setenv(MAPPING_DOWNLOAD_DIR_ENV_VAR, d)
         yield d
 
 
@@ -20,6 +20,7 @@ def patch_conda_mapping_fixes(mocker, conda_mapping_fixes):
 
 class TestMapping:
     @pytest.mark.parametrize("conda_mapping", [{"pytest": "pytest-conda", "other": "other-conda"}, dict()])
+    @pytest.mark.parametrize("mapping_url", [None, "https://example.com/mapping.yaml"])
     @pytest.mark.parametrize(
         "conda_mapping_fixes",
         [{"corrected-mapping": "corrected"}, {"pytest": "pytest-corrected"}],
@@ -32,16 +33,27 @@ class TestMapping:
         patch_conda_mapping_fixes,
         conda_mapping_fixes,
         mocked_responses,
+        mapping_url,
+        monkeypatch,
     ):
         """
         Test project conda_mapping downloads conda mapping just one and mapping is as expected
         """
-        from pdm_conda.mapping import MAPPINGS_URL, get_conda_mapping, get_pypi_mapping
+        from pdm_conda.mapping import (
+            MAPPING_URL,
+            MAPPING_URL_ENV_VAR,
+            get_conda_mapping,
+            get_pypi_mapping,
+        )
 
         get_pypi_mapping.cache_clear()
         get_conda_mapping.cache_clear()
+        if mapping_url is None:
+            mapping_url = MAPPING_URL
+        monkeypatch.setenv(MAPPING_URL_ENV_VAR, mapping_url)
 
         assert project.config["conda.pypi-mapping.download-dir"] == patch_download_dir
+        assert project.config["conda.pypi-mapping.url"] == mapping_url
 
         response = ""
         for pypi_name, conda_name in conda_mapping.items():
@@ -52,7 +64,7 @@ class TestMapping:
                 mapping_source: other
                 pypi_name: {pypi_name}
             """
-        rsp = mocked_responses.get(MAPPINGS_URL, body=response)
+        rsp = mocked_responses.get(mapping_url, body=response)
 
         corrected_mapping = dict(conda_mapping)
         corrected_mapping.update(conda_mapping_fixes)
@@ -76,6 +88,7 @@ class TestMapping:
         conda_mapping_fixes,
         package,
         mocked_responses,
+        monkeypatch,
     ):
         self.test_download_mapping(
             patch_download_dir,
@@ -84,6 +97,8 @@ class TestMapping:
             patch_conda_mapping_fixes,
             conda_mapping_fixes,
             mocked_responses,
+            None,
+            monkeypatch,
         )
         from pdm_conda.mapping import pypi_to_conda